Move the `map` function definition to `idna.cpp`
and move the internal header files: `idna_table.h` and `iterate_utf.h`
to the `src/` directory.
