+++ tools/amalgamate/config-cpp.json
@@ -2,6 +2,8 @@
   "project": "upa/idna.cpp",
   "target": "single_include/upa/idna.cpp",
   "sources": [
+    "src/idna_table.h",
+    "src/iterate_utf.h",
     "src/idna.cpp",
     "src/idna_table.cpp",
     "src/nfc_table.h",
